Сorrected remarks

+++ Uspihh.Models/DTOModels/GroupDTO.cs
@@ -1,8 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace Uspihh.Models.DTOModels
 {
@@ -10,7 +6,7 @@ namespace Uspihh.Models.DTOModels
     {
         public long GroupId { get; set; }
         public string GroupName { get; set; }
-        public int MaxCountOfStudent { get; set; }
+        public short MaxCountOfStudent { get; set; }
         public DateTime StartDate { get; set; }
         public DateTime EndDate { get; set; }
         public int MounthPrice { get; set; }
