+++ Uspihh.Models/EntityModels/GroupEntity.cs
@@ -6,7 +6,7 @@ namespace Uspihh.Models.EntityModels
     {
         public long GroupId { get; set; }
         public string GroupName { get; set; }
-        public int MaxCountOfStudent { get; set; }
+        public short MaxCountOfStudent { get; set; }
         public DateTime StartDate { get; set; }
         public DateTime EndDate { get; set; }
         public int MounthPrice { get; set; }
